Implement GPU acceleration for canvases
