+++ visualizer/src/Canvas/Labeled/LabeledCanvas.js
@@ -1,104 +1,95 @@
 import { useSelector } from '@xstate/react';
-import React, { useEffect, useRef } from 'react';
+import colormap from 'colormap';
+import { GPU } from 'gpu.js';
+import { useEffect, useMemo, useRef } from 'react';
 import { useCanvas, useFeature, useLabeled, useSelect } from '../../ProjectContext';
-import { highlightImageData, opacityImageData } from '../canvasUtils';
 
-export const LabeledCanvas = ({ className }) => {
+function getColormap(colormapName, nShades, format) {
+  const minimum = colormap.cm[colormapName].length;
+  let gap = minimum;
+  for (let i = 0; gap < minimum; gap++) {
+    if (i * (nShades - 1) + nShades >= minimum) {
+      gap = i;
+      break;
+    }
+  }
+  const tempNShades = Math.max(minimum, nShades + gap * (nShades - 1));
+  const tempCmap = colormap({ colormap: colormapName, nShades: tempNShades, format });
+  return tempCmap.filter((_, i) => i % gap === 0);
+}
+
+const highlightColor = [255, 0, 0];
+
+export const LabeledCanvas = ({ setCanvases }) => {
   const canvas = useCanvas();
-  const sx = useSelector(canvas, (state) => state.context.sx);
-  const sy = useSelector(canvas, (state) => state.context.sy);
-  const zoom = useSelector(canvas, (state) => state.context.zoom);
-  const scale = useSelector(canvas, (state) => state.context.scale);
-  const sw = useSelector(canvas, (state) => state.context.width);
-  const sh = useSelector(canvas, (state) => state.context.height);
-
-  const width = sw * scale * window.devicePixelRatio;
-  const height = sh * scale * window.devicePixelRatio;
+  const width = useSelector(canvas, (state) => state.context.width);
+  const height = useSelector(canvas, (state) => state.context.height);
 
   const labeled = useLabeled();
   const featureIndex = useSelector(labeled, (state) => state.context.feature);
   const highlight = useSelector(labeled, (state) => state.context.highlight);
   const opacity = useSelector(labeled, (state) => state.context.opacity);
+  // const highlightColor = [255, 0, 0];
 
   const feature = useFeature(featureIndex);
-  const labeledImage = useSelector(feature, (state) => state.context.labeledImage);
+  // const labeledImage = useSelector(feature, (state) => state.context.labeledImage);
+  // const cmap = useSelector(feature, (state) => state.context.colormap);
   let labeledArray = useSelector(feature, (state) => state.context.labeledArray);
   if (!labeledArray) {
-    labeledArray = Array(sh).fill(Array(sw).fill(0));
+    labeledArray = Array(height * width).fill(0);
   }
 
   const select = useSelect();
   const foreground = useSelector(select, (state) => state.context.foreground);
-  const background = useSelector(select, (state) => state.context.background);
+  // const background = useSelector(select, (state) => state.context.background);
+  const maxLabel = useSelector(select, (state) => Math.max(...Object.keys(state.context.labels)));
+  const cmap = useMemo(
+    () =>
+      colormap({
+        colormap: 'viridis',
+        nshades: Math.max(9, maxLabel),
+        format: 'rgba',
+      }),
+    [maxLabel]
+  );
 
+  const kernel = useRef();
   const canvasRef = useRef();
-  const ctx = useRef();
-  const hiddenCanvasRef = useRef();
-  const hiddenCtx = useRef();
 
   useEffect(() => {
-    ctx.current = canvasRef.current.getContext('2d');
-    ctx.current.imageSmoothingEnabled = false;
+    const gpu = new GPU();
+    kernel.current = gpu
+      .createKernel(function (labels, colormap, foreground, highlight, highlightColor, opacity) {
+        const n = this.thread.x + this.constants.w * (this.constants.h - this.thread.y);
+        const label = labels[n];
+        if (highlight && label === foreground && foreground !== 0) {
+          const [r, g, b] = highlightColor;
+          this.color(r / 255, g / 255, b / 255, opacity);
+        } else if (label !== 0) {
+          const [r, g, b] = colormap[label - 1];
+          this.color(r / 255, g / 255, b / 255, opacity);
+        } else {
+          this.color(0, 0, 0, opacity);
+        }
+      })
+      .setConstants({ w: width, h: height })
+      .setOutput([width, height])
+      .setGraphical(true)
+      .setDynamicArguments(true);
+    canvasRef.current = kernel.current.canvas;
   }, [width, height]);
 
   useEffect(() => {
-    hiddenCtx.current = hiddenCanvasRef.current.getContext('2d');
-  }, [sw, sh]);
+    kernel.current(labeledArray, cmap, foreground, highlight, highlightColor, opacity);
+    setCanvases((canvases) => ({ ...canvases, labeled: canvasRef.current }));
+    // return () =>
+    //   setCanvases((canvases) => {
+    //     delete canvases['labeled'];
+    //     return { ...canvases };
+    //   });
+  }, [labeledArray, cmap, foreground, highlight, opacity, setCanvases]);
 
-  useEffect(() => {
-    hiddenCtx.current.drawImage(labeledImage, 0, 0);
-    let data = hiddenCtx.current.getImageData(0, 0, sw, sh);
-    if (highlight && foreground !== 0) {
-      const red = [255, 0, 0, 255];
-      highlightImageData(data, labeledArray, foreground, red);
-    }
-    opacityImageData(data, opacity);
-    hiddenCtx.current.putImageData(data, 0, 0);
-  }, [labeledImage, labeledArray, foreground, highlight, opacity, sh, sw]);
-
-  useEffect(() => {
-    ctx.current.save();
-    ctx.current.clearRect(0, 0, width, height);
-    ctx.current.drawImage(
-      hiddenCanvasRef.current,
-      sx,
-      sy,
-      sw / zoom,
-      sh / zoom,
-      0,
-      0,
-      width,
-      height
-    );
-    ctx.current.restore();
-  }, [
-    labeledImage,
-    labeledArray,
-    foreground,
-    highlight,
-    opacity,
-    sw,
-    sh,
-    sx,
-    sy,
-    zoom,
-    width,
-    height,
-  ]);
-
-  return (
-    <>
-      {/* hidden processing canvas */}
-      <canvas id='labeled-processing' hidden={true} ref={hiddenCanvasRef} width={sw} height={sh} />
-      <canvas
-        id='labeled-canvas'
-        ref={canvasRef}
-        width={width}
-        height={height}
-        className={className}
-      />
-    </>
-  );
+  return null;
 };
 
 export default LabeledCanvas;
